+++ vite-plugin/vite-plugin-appsscript.js
@@ -1,26 +1,94 @@
+/** @module GoogleAppsScriptExportsPlugin */
+
 import fs from 'fs';
-import path from 'path';
+import { resolve } from 'path';
+import { parse } from '@babel/parser';
+import t from '@babel/traverse';
+import extractExports from './traverse/extract-exports.js';
+
+const traverse = t.default;
 
 /**
- * Exports the initianton function from the library for better semantics
- * @param {string} filePath - The path to the code to process
- * @param {string} propName - The name of the prperty to extract from the IIFE
- * @returns {object} - Rollup plugin object
+ * Extracts the exports as stand-alone functions
+ * @param {string} origPath The file to extract the exports from
+ * @param {string} exportsPath The file to extract the exports to
+ * @param {string} varName The name of the variable that stores the return value of the IIFE
+ * @param {GasExportOptions} [options] The options for the plugin
  */
-export function AppsScriptPlugin(filePath, propName) {
+export async function GoogleAppsScriptExportsPlugin(
+  origPath = 'dist/gas/server/server.iife.js',
+  exportsPath = 'dist/gas/exports.js',
+  varName = 'lib_',
+  options = {
+    exportsFile: true,
+    copyFiles: [{ from: './appsscript.json', to: 'dist/gas/appsscript.json' }],
+  }
+) {
   return {
-    name: 'vite-plugin-appscript',
+    name: 'vite-plugin-appscript-exports',
     async closeBundle() {
-      setTimeout(() => {
-        filePath = path.join(process.cwd(), filePath);
-        let data = fs.readFileSync(filePath, 'utf8');
-        data += `\n${propName} = ${propName}.${propName};`;
-        fs.mkdirSync(filePath.split('/').slice(0, -1).join('/'), {
-          recursive: true,
-        });
-        fs.writeFileSync(filePath, data, 'utf8');
-        fs.copyFileSync('appsscript.json', 'dist/appsscript.json');
-      }, 0);
+      try {
+        console.log('closeBundle');
+        const code = fs.readFileSync(origPath, 'utf-8');
+        const ast = parse(code);
+        const exportDetails = extractExports(ast);
+        // fs.writeFileSync(
+        //   'exportDetails.json',
+        //   JSON.stringify(exportDetails, null, 2)
+        // );
+
+        const getParam = paramObj => {
+          if (!paramObj.defaultValue) return paramObj.name;
+          if ('string' === typeof paramObj.defaultValue)
+            return `${paramObj.name} = '${paramObj.defaultValue.replace(
+              /'/g,
+              "\\'"
+            )}'`;
+          return `${paramObj.name} = ${paramObj.defaultValue}`;
+        };
+        const exportsText =
+          '/* eslint-disable no-undef */\n' +
+          exportDetails
+            .map(fn => {
+              return null === fn.parameters
+                ? `const ${fn.name} = ${varName}.${fn.name};`
+                : [
+                    `function ${fn.name}(${fn.parameters
+                      .map(getParam)
+                      .join(', ')}) {`,
+                    `  return ${varName}.${fn.name}(${fn.parameters
+                      .map(p => p.name)
+                      .join(', ')});`,
+                    '}',
+                  ].join('\n');
+            })
+            .join('\n');
+
+        fs.writeFileSync(exportsPath, exportsText, 'utf8');
+
+        if (false === options.exportsFile) {
+          const exp = fs.readFileSync(
+            resolve(process.cwd(), exportsPath),
+            'utf8'
+          );
+          fs.appendFileSync(
+            resolve(process.cwd(), origPath),
+            `\n${exp}`,
+            'utf8'
+          );
+          fs.unlinkSync(resolve(process.cwd(), exportsPath));
+        }
+      } catch (err) {
+        console.error(`Could not create exports: ${err}`);
+      }
+
+      if (!options.copyFiles || !options.copyFiles.length) return;
+      options.copyFiles.forEach(file => {
+        fs.copyFileSync(
+          resolve(process.cwd(), file.from),
+          resolve(process.cwd(), file.to)
+        );
+      });
     },
   };
 }